// frontend/src/client/types.gen.ts
export type Body_login_login_access_token = {
    grant_type?: (string | null);
    username: string;
    password: string;
    scope?: string;
    client_id?: (string | null);
    client_secret?: (string | null);
};
// ...
export type UserPublic = {
    email: string;
    is_active?: boolean;
    is_superuser?: boolean;
    full_name?: (string | null);
    birthdate?: (string | null);
    phone_number?: (string | null);
    role?: string;
    company_id?: (string | null);
    companies?: string;
    last_login_at?: (string | null);
    onboarding_completed?: boolean;
    email_verified?: boolean;
    email_verified_at?: (string | null);
    has_completed_tour?: boolean;
    completed_tour_at?: (string | null);
    id: string;
    created_at?: (string | null);
    plan?: (string | null);
    billing_period?: (string | null);
};

export type UserRegister = {
    email: string;
    password: string;
    full_name?: (string | null);
    birthdate?: (string | null);
    phone_number?: (string | null);
};